+++ src/index.js
@@ -86,6 +86,7 @@ export { default as TileEquipo } from './components/TileEquipo.jsx'
 export { default as Stepper } from './components/Stepper.jsx'
 export { default as CodigoQr } from './components/CodigoQr.jsx'
 export { default as FichaCertificado } from './components/FichaCertificado.jsx'
+export { default as VistaPreviaPapel, ANCHOS_PAPEL } from './components/VistaPreviaPapel.jsx'
 export { QR_OPCIONES, qrDataUrl } from './utils/qr.js'
 export {
   ESTADOS_ITEM,